+++ packages/tempus-client_v2/src/components/wallet/Wallet.tsx
@@ -15,7 +15,7 @@ import { WalletContext } from '../../context/walletContext';
 import NumberUtils from '../../services/NumberUtils';
 import UserWallet from '../../interfaces/UserWallet';
 import getText from '../../localisation/getText';
-import useENS from '../../utils/useENS';
+import useENS from '../../hooks/useENS';
 import shortenAccount from '../../utils/shortenAccount';
 import getStorageService from '../../services/getStorageService';
 import getNotificationService from '../../services/getNotificationService';
